implemented input validation for schedule creation
the program will keep asking for a valid input until one is obtained. .nextInt() did not work when trying to catch the exception, as it was stuck in an infinite loop. May be due to a issue with the scope of .nextInt exceptions. Taking input as a string then parsing it to an int worked.